requirements.txt
# Core web
flask>=3.1.2
flask-socketio>=5.5.1
python-socketio>=5.14.2

# AI providers and clients
openai>=1.54.0
httpx>=0.27.2

# Data & utils
requests>=2.32.3
pyyaml>=6.0.2
pandas>=2.3.3
numpy>=1.26.0

# Supabase client
supabase>=2.22.0

# NLP / embeddings (used by verifier & matcher)
sentence-transformers>=2.7.0

# Security / crypto / DB
cryptography>=43.0.0
psycopg2-binary>=2.9.9

# Google APIs (optional: Drive & OAuth)
google-generativeai>=0.7.2
google-api-python-client>=2.151.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0

# Env convenience (optional)
python-dotenv>=1.0.1
